requirements.txt
alembic==1.10.2
anyio==3.6.2
APScheduler==3.9.1.post1
autopep8==2.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
pycodestyle==2.10.0
PyMySQL==1.0.2
python-telegram-bot==20.0a4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.5.post1
tomli==2.0.1
tornado==6.2
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.15.0
